requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0

# LangChain and AI - GEMINI ONLY
langchain==0.1.8
langchain-community==0.0.14
langchain-google-genai==0.0.6
google-generativeai==0.3.0

# Language and NLP
langdetect==1.0.9
transformers==4.36.0
torch==2.1.1
numpy<2
scikit-learn==1.3.2

# Web Search and Tools
tavily-python==0.2.0
google-search-results==2.4.2
requests==2.31.0

# Vector Database
chromadb==0.4.21
sentence-transformers==2.2.2

# Server
waitress==2.1.2
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2

# Monitoring and Logging
python-json-logger==2.0.7